# merchants/docs/swagger3.json
{"openapi":"3.0.1","info":{"contact":{}},"servers":[{"url":"/"}],"paths":{"/api/v1/core/merchants":{"get":{"tags":["Merchants"],"summary":"Get merchants","description":"Get merchant","responses":{"200":{"description":"Success Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/rest.merchantsResult"}}}},"500":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/errorDomain.SmartError"}}}}},"security":[{"BearerAuth":[]}]},"post":{"tags":["Merchants"],"summary":"Create merchant","description":"Create merchant","requestBody":{"description":"Create merchant body","content":{"application/json":{"schema":{"$ref":"#/components/schemas/domain.CreateMerchantBody"}}},"required":true},"responses":{"200":{"description":"Success Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/httpResponse.IdResult"}}}},"500":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/errorDomain.SmartError"}}}}},"security":[{"BearerAuth":[]}],"x-codegen-request-body-name":"createMerchantBody"}},"/api/v1/core/merchants/{merchantId}":{"put":{"tags":["Merchants"],"summary":"Update merchant","description":"Update merchant","parameters":[{"name":"merchantId","in":"path","description":"merchant id","required":true,"schema":{"type":"string"}}],"requestBody":{"description":"Update merchant body","content":{"application/json":{"schema":{"$ref":"#/components/schemas/domain.UpdateMerchantBody"}}},"required":true},"responses":{"200":{"description":"Success Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/httpResponse.StatusResult"}}}},"500":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/errorDomain.SmartError"}}}}},"security":[{"BearerAuth":[]}],"x-codegen-request-body-name":"updateMerchantBody"},"delete":{"tags":["Merchants"],"summary":"Delete a merchant","description":"Delete merchant","parameters":[{"name":"merchantId","in":"path","description":"merchant id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Success Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/rest.deleteMerchantsResult"}}}},"500":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/errorDomain.SmartError"}}}}},"security":[{"BearerAuth":[]}]}}},"components":{"schemas":{"domain.CreateMerchantBody":{"required":["address","description","document","image_path","industry","name","phone"],"type":"object","properties":{"address":{"type":"string","description":"Description: the address of the merchant","example":"123 Main Street"},"description":{"type":"string","description":"Description: the description of the merchant","example":"Proveedor de servicios de mantenimiento"},"document":{"type":"string","description":"Description: the document of the merchant","example":"123456789"},"image_path":{"type":"string","description":"Description: the image_path of the merchant","example":"https://example.com/images/odin_logo.png"},"industry":{"type":"string","description":"Description: the industry of the merchant","example":"Mantenimiento"},"name":{"type":"string","description":"Description: the name of the merchant","example":"Odin Corp"},"phone":{"type":"string","description":"Description: the phone of the merchant","example":"+1234567890"}}},"domain.Merchant":{"required":["address","created_at","description","document","id","image_path","industry","name","phone"],"type":"object","properties":{"address":{"type":"string","description":"Description: the address of the merchant","example":"123 Main Street"},"created_at":{"type":"string","description":"Description: the created_at of the merchant","example":"2023-11-10 08:10:00"},"description":{"type":"string","description":"Description: the description of the merchant","example":"Proveedor de servicios de mantenimiento"},"document":{"type":"string","description":"Description: the document of the merchant","example":"123456789"},"id":{"type":"string","description":"Description: the id of the merchant","example":"739bbbc9-7e93-11ee-89fd-0242ac110016"},"image_path":{"type":"string","description":"Description: the image_path of the merchant","example":"https://example.com/images/odin_logo.png"},"industry":{"type":"string","description":"Description: the industry of the merchant","example":"Mantenimiento"},"name":{"type":"string","description":"Description: the name of the merchant","example":"Odin Corp"},"phone":{"type":"string","description":"Description: the phone of the merchant","example":"+1234567890"}}},"domain.PaginationResults":{"required":["current_page","last_page","size_page","total"],"type":"object","properties":{"current_page":{"type":"integer"},"from":{"type":"integer"},"last_page":{"type":"integer"},"size_page":{"type":"integer"},"to":{"type":"integer"},"total":{"type":"integer"}}},"domain.UpdateMerchantBody":{"required":["address","description","document","image_path","industry","name","phone"],"type":"object","properties":{"address":{"type":"string","description":"Description: the address of the merchant","example":"123 Main Street"},"description":{"type":"string","description":"Description: the description of the merchant","example":"Proveedor de servicios de mantenimiento"},"document":{"type":"string","description":"Description: the document of the merchant","example":"123456789"},"image_path":{"type":"string","description":"Description: the image_path of the merchant","example":"https://example.com/images/odin_logo.png"},"industry":{"type":"string","description":"Description: the industry of the merchant","example":"Mantenimiento"},"name":{"type":"string","description":"Description: the name of the merchant","example":"Odin Corp"},"phone":{"type":"string","description":"Description: the phone of the merchant","example":"+1234567890"}}},"errorDomain.LayerErr":{"type":"string","enum":["domain","infrastructure","interface","use_case"],"x-enum-varnames":["Domain","Infra","Interface","UseCase"]},"errorDomain.LevelErr":{"type":"string","enum":["info","warning","error","fatal"],"x-enum-varnames":["LevelInfo","LevelWarning","LevelError","LevelFatal"]},"errorDomain.SmartError":{"type":"object","properties":{"code":{"type":"string"},"description":{"type":"string"},"error":{"type":"object"},"function":{"type":"string"},"httpStatus":{"type":"integer"},"layer":{"$ref":"#/components/schemas/errorDomain.LayerErr"},"level":{"$ref":"#/components/schemas/errorDomain.LevelErr"},"messages":{"type":"array","items":{"type":"string"}},"raw":{"type":"string"}}},"httpResponse.IdResult":{"required":["data","status"],"type":"object","properties":{"data":{"type":"string","example":"201"},"status":{"type":"integer"}}},"httpResponse.StatusResult":{"required":["status"],"type":"object","properties":{"status":{"type":"integer","example":200}}},"rest.deleteMerchantsResult":{"required":["data","status"],"type":"object","properties":{"data":{"type":"boolean"},"status":{"type":"integer"}}},"rest.merchantsResult":{"required":["data","pagination","status"],"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/domain.Merchant"}},"pagination":{"$ref":"#/components/schemas/domain.PaginationResults"},"status":{"type":"integer"}}}},"securitySchemes":{"BearerAuth":{"type":"apiKey","name":"Authorization","in":"header"}}}}